defconfig: Don't sync when entering suspend
Syncing can take a long time in some cases, and the risk of us losing
power or crashing while in suspend on a battery-backed device is very#
Please enter the commit message for your changes. Lines starting low.
Android also enters and exits suspend very frequently, which# with '#'
will be ignored, and an empty message aborts the commit.
exacerbates the cost.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>